+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 - Add P2p feature on videojs player
 - Webtorrent tracker for P2P video feature
 - eslint sort the modules name (#2338)
+- Add retention date and s3 lifecycle rules to classroom / video
 
 ### Changed